requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
boto3==1.24.54
botocore==1.27.96
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
execnet==2.0.2
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
gspread==5.4.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpickle==2.0.0
mail-parser==3.15.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.7.0.68
openpyxl==3.0.9
orjson==3.10.6
packaging==23.2
pact-python==2.2.0
percy-selenium==1.0.4
pluggy==1.4.0
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyGithub==1.55
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==0.9.3
PyNaCl==1.5.0
PyPika==0.48.8
pytest==7.2.1
pytest-instafail==0.4.2
pytest-rerunfailures==10.2
pytest-xdist==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2021.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
s3transfer==0.6.2
selenium==3.141.0
setuptools==69.1.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
slackclient==2.9.3
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.7
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
webdriver-manager==4.0.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
